# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraSol API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/counties/{id}": {
            "get": {
                "tags": [
                    "Counties"
                ],
                "summary": "Get county information",
                "description": "Returns county data",
                "operationId": "getCountyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "County id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Counties"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Counties"
                ],
                "summary": "Update existing county",
                "description": "Returns updated county data",
                "operationId": "updateCounties",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "County id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CountiesFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Counties"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Counties"
                ],
                "summary": "Delete existing county",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCounty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "County id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CountiesFormRequest": {
                "title": "County Form request",
                "required": [
                    "county_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "county_name": {
                        "description": "County Name",
                        "type": "string",
                        "example": "Nairobi"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CountiesFormRequest"
                }
            },
...
            "Counties": {
                "required": [
                    "county_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "county_name": {
                        "description": "County Name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Counties"
                }
            },
...
        }
    }
}